Read user config after creating socket and fifo
My guile-daemon config file (which wasn't touched for many months)
suddenly stopped working (guile-daemon simply freezes on reading config
file without any error) and this appeared to be a fix for this problem.
I have no idea what caused it, maybe update of guile or some of the
underlying libraries.  Anyway, a fix is a fix.

* modules/daemon.scm: Load config file after starting a server and
  making fifo file.